added materials, something is wrong with tri intersection or the scene setup, all rays intersect the hexagons

+++ include/Image.h
@@ -5,9 +5,9 @@ class Image
 public:
 	struct Color
 	{
-		double* r;
-		double* g;
-		double* b;
+		float* r;
+		float* g;
+		float* b;
 	};
 
 	Image();
